update refs to app redux submodule
