+++ src/redux/app/thunks/index.ts
@@ -1,0 +1,1 @@
+export * from "./getReadOnlyMode";
